+++ dgisim/src/character/character.py
@@ -39,6 +39,7 @@ __all__ = [
     "Mona",
     "RhodeiaOfLoch",
     "Tighnari",
+    "Xingqiu",
 ]
 
 
@@ -1357,6 +1358,95 @@ class Tighnari(Character):
 
     @classmethod
     def from_default(cls, id: int = -1) -> Tighnari:
+        return cls(
+            id=id,
+            alive=True,
+            hp=10,
+            max_hp=10,
+            energy=0,
+            max_energy=2,
+            hiddens=stts.Statuses(()),
+            equipments=stts.EquipmentStatuses(()),
+            statuses=stts.Statuses(()),
+            elemental_aura=ElementalAura.from_default(),
+        )
+
+
+class Xingqiu(Character):
+    _ELEMENT = Element.HYDRO
+    _WEAPON_TYPE = WeaponType.SWORD
+    _TALENT_STATUS = stt.TheScentRemainedStatus
+
+    @override
+    @classmethod
+    def skill_cost(cls, skill_type: CharacterSkill) -> AbstractDices:
+        if skill_type is CharacterSkill.NORMAL_ATTACK:
+            return AbstractDices({
+                Element.HYDRO: 1,
+                Element.ANY: 2,
+            })
+        elif skill_type is CharacterSkill.ELEMENTAL_SKILL1:
+            return AbstractDices({
+                Element.HYDRO: 3,
+            })
+        elif skill_type is CharacterSkill.ELEMENTAL_BURST:
+            return AbstractDices({
+                Element.HYDRO: 3,
+            })
+        raise Exception("Not Reached!")
+
+    def _normal_attack(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return normal_attack_template(
+            game_state=game_state,
+            source=source,
+            element=Element.PHYSICAL,
+            damage=2,
+        )
+
+    def _elemental_skill1(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.HYDRO,
+                damage=2,
+                damage_type=DamageType(elemental_skill=True),
+            ),
+            eft.ApplyElementalAuraEffect(
+                target=source,
+                element=Element.HYDRO,
+            ),
+            eft.UpdateCombatStatusEffect(
+                target_pid=source.pid,
+                status=stt.RainSwordStatus(usages=case_val(self.talent_equiped(), 3, 2))
+            ),
+        )
+
+    def _elemental_burst(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.EnergyDrainEffect(
+                target=source,
+                drain=self.get_max_energy(),
+            ),
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.HYDRO,
+                damage=1,
+                damage_type=DamageType(elemental_burst=True),
+            ),
+            eft.ApplyElementalAuraEffect(
+                target=source,
+                element=Element.HYDRO,
+            ),
+            eft.AddCombatStatusEffect(
+                target_pid=source.pid,
+                status=stt.RainbowBladeworkStatus,
+            ),
+        )
+
+    @classmethod
+    def from_default(cls, id: int = -1) -> Self:
         return cls(
             id=id,
             alive=True,