+++ dgisim/tests/test_characters/test_arataki_itto.py
@@ -1,4 +1,3 @@
-
 import unittest
 
 from dgisim.tests.test_characters.common_imports import *